% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatin_potential_preparation.R
\name{chromatin_potential_prepare}
\alias{chromatin_potential_prepare}
\title{Prepare chromatin potential}
\usage{
chromatin_potential_prepare(
  mat_x,
  mat_y,
  df_x,
  df_y,
  vec_start,
  list_end,
  dim_method = "pca",
  nn_method = "annoy",
  form_method = "average",
  est_method = "glmnet",
  cand_method = "nn_any",
  rec_method = "distant_cor",
  options = list(),
  verbose = T
)
}
\arguments{
\item{mat_x}{full data for Modality 1, where each row is a cell and each column is a variable}

\item{mat_y}{full data for Modality 2, where each row is a cell and each column is a variable}

\item{df_x}{the data frame containing information of Modality 1}

\item{df_y}{the data frame containing information of Modality 2}

\item{vec_start}{integers between 1 and \code{nrow(mat_x)} to denote the cells at the start state}

\item{list_end}{list of integers between 1 and \code{nrow(mat_x)} to denote the cells any of the end states}

\item{dim_method}{string}

\item{nn_method}{string}

\item{form_method}{string}

\item{est_method}{string}

\item{cand_method}{string}

\item{rec_method}{string}

\item{options}{list}

\item{verbose}{boolean}
}
\value{
object of class \code{chromatin_potential_prep}
}
\description{
To use this function properly, we assume that assumes that \code{mat_x} and 
\code{mat_y} are normalized
}
