% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_method.R
\name{.candidate_set}
\alias{.candidate_set}
\title{Find candidate cells}
\usage{
.candidate_set(mat_x, mat_y, df_res, nn_mat, cand_options)
}
\arguments{
\item{mat_x}{full data for Modality 1, where each row is a cell and each column is a variable}

\item{mat_y}{full data for Modality 2, where each row is a cell and each column is a variable}

\item{df_res}{data frame recording the current results, generated within \code{chromatin_potential}}

\item{nn_mat}{the nearest-neighbor matrix, output from \code{chromatin_potential_prepare}}

\item{cand_options}{one of the outputs from \code{.chrom_options}}
}
\value{
a list containing \code{vec_cand} (vector of integers between 1 and \code{nrow(mat_x)})
and a list \code{diagnostic} for possible diagnostics
}
\description{
Find other cells based on \code{mat_x} (the data for Modality 1),
\code{mat_y} (the data for Modality 2), and
\code{df_res} to potentially be recruited in this current
iteration.
}
\details{
The options are:
\itemize{
\item \code{nn_any}: Find the candidates based only on which
unrecruited cells are one of the nearest neighbors (based on \code{nn_mat})
to either any of the previously-recruited cells in the latest iteration
or in any of the previous iterations (depending on \code{cand_options$only_latest})
\item \code{nn_freq}: Find the \code{cand_options$num_cand} candidates 
which are the \code{cand_options$num_cand} unrecruited cells 
which have the most nearest neighbors (based on \code{nn_mat})
to any previously-recruited cells
\item \code{all}: All the unrecruited cells are considered as candidates
}
}
