% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_method.R
\name{.estimate_g}
\alias{.estimate_g}
\title{Estimate the GLM that links the two modalities}
\usage{
.estimate_g(mat_x1, mat_y2, est_options)
}
\arguments{
\item{mat_x1}{Output of \code{.init_est_matrices} or \code{.update_estimation_matrices}, representing
the data for Modality 1}

\item{mat_y2}{Output of \code{.init_est_matrices} or \code{.update_estimation_matrices}, representing
the data for Modality 2}

\item{est_options}{one of the outputs from \code{.chrom_options}}
}
\value{
a list of a matrix \code{mat_g} (of dimensions \code{ncol(mat_x1)} by \code{ncol(mat_y2)})
and a vector \code{vec_g} (of length \code{ncol(mat_y2)})
}
\description{
Estimate the GLM for each variable of \code{mat_y2} onto
all the variables in \code{mat_x1}.
}
\details{
If \code{est_options$enforce_cis} is \code{TRUE}, this function
will assume \code{est_options$ht_map} is a code{hash} object
already populated with all the variables in Modality 1 that 
are associated with Modality 2

The options are:
\itemize{
\item \code{glmnet}: Estimate the link from \code{mat_x1} to \code{mat_y2}
using \code{glmnet::glmnet}. The options \code{est_options$family}, \code{est_options$alpha}, 
\code{est_options$standardize}, and \code{est_options$intercept} are the arguments
for \code{glmnet::glmnet}. If \code{est_options$cv} is \code{TRUE}, then
\code{glmnet::cv.glmnet} is used to select the regularization parameter
using \code{est_options$nfolds} folds via \code{est_options$cv_choice}.
If \code{est_options$cv} is \code{FALSE}, use the densest solution.
\code{est_options$switch} is a boolean that automatically switches 
from \code{glmnet::glmnet} to \code{stats::glm} if there are 
\code{est_options$switch_cutoff} times more cells in \code{mat_x1} than variables
in \code{mat_x1}. 
}
}
