% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forming_method.R
\name{.init_est_matrices}
\alias{.init_est_matrices}
\title{Initialize the matrices for estimation later}
\usage{
.init_est_matrices(mat_x, mat_y, vec_start, list_end)
}
\arguments{
\item{mat_x}{full data for Modality 1, where each row is a cell and each column is a variable}

\item{mat_y}{full data for Modality 2, where each row is a cell and each column is a variable}

\item{vec_start}{integers between 0 and \code{nrow(mat_x)} to denote the cells at the start state}

\item{list_end}{integers between 0 and \code{nrow(mat_x)} to denote the cells any of the end states}
}
\value{
list of 3 matrices, \code{mat_x1} and \code{mat_y1} and \code{mat_y2}, as
well as vector of indices \code{idx1}
}
\description{
Initialize the matrices for estimation later
}
