% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forming_method.R
\name{.update_estimation_matrices}
\alias{.update_estimation_matrices}
\title{Update the matrices for estimation later}
\usage{
.update_estimation_matrices(
  mat_x,
  mat_y,
  mat_x1,
  mat_y1,
  mat_y2,
  idx1,
  rec,
  form_options
)
}
\arguments{
\item{mat_x}{full data for Modality 1, where each row is a cell and each column is a variable}

\item{mat_y}{full data for Modality 2, where each row is a cell and each column is a variable}

\item{mat_x1}{previous \code{mat_x1}, say, from an earlier call to \code{.init_est_matrices} or \code{.update_estimation_matrices}}

\item{mat_y1}{previous \code{mat_y1}, say, from an earlier call to \code{.init_est_matrices} or \code{.update_estimation_matrices}}

\item{mat_y2}{previous \code{mat_y2}, say, from an earlier call to \code{.init_est_matrices} or \code{.update_estimation_matrices}}

\item{idx1}{previous \code{idx1}, say, from an earlier call to \code{.init_est_matrices} or \code{.update_estimation_matrices}}

\item{rec}{output from \code{.recruit_next}}

\item{form_options}{one of the outputs from \code{.chrom_options}}
}
\value{
list of 3 matrices, \code{mat_x1} and \code{mat_y1} and \code{mat_y2}, as
well as vector of indices \code{idx1}
}
\description{
Using the output of \code{.recruit_next} (in \code{rec}), update
\code{mat_x1}, \code{mat_y1}, \code{mat_y2}, \code{idx1} by grabbing
the appropriate cells in \code{mat_x} and \code{mat_y} based on the
options in \code{form_options}
}
\details{
The options are:
\itemize{
\item \code{literal}: If \code{rec$list_to[[i]]} matches many cells from a single cell in \code{rec$vec_from[i]},
form one row in \code{mat_y2} for each matched cell in \code{rec$list_to[[i]]} (and 
duplicate the cell in \code{rec$vec_from[i]} in \code{mat_x1} as many times as needed)
\item \code{average}: If \code{rec$list_to[[i]]} matches many cells from a single cell in \code{rec$vec_from[i]},
average all the variables among the matched cell in \code{rec$list_to[[i]]} based on
\code{form_options$average} to form one row in \code{mat_y2} (and one row in 
\code{mat_x1} for \code{rec$vec_from[i]})
}
}
