% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generating data}
\usage{
generate_data(
  obj_next,
  max_n = 2 * length(obj_next$ht),
  number_runs = 5,
  sample_perc = 1,
  time_tol = 0.01,
  verbose = T
)
}
\arguments{
\item{obj_next}{returned by \code{prepare_obj_nextcell}}

\item{max_n}{positive integer, for maximum number of cells for each of the \code{number_runs} runs}

\item{number_runs}{positive integer, for number of times we start at \code{obj_next$vec_startx} and follow that cell's trajectory}

\item{sample_perc}{number between \code{0} and \code{1}, denoting what percentage of cells we've generated in this simulation
are returned in the final output. Here, \code{1} means all the cells generated are returned.}

\item{time_tol}{positive integer between  \code{0} and \code{1}, typically close to \code{0},
denoting the stopping criteria for each of the \code{number_runs} runs. A run
stops when it generates a cell whose pseudotime (as dictated in \code{obj_next})
is larger than \code{1-time_tol}.}

\item{verbose}{boolean}
}
\value{
object of class \code{mf_simul} with the following components: 
\code{df_x} (the data frame from \code{obj_next} containing information of Modality 1),
\code{df_y} (the data frame from \code{obj_next} containing information of Modality 2),
\code{obs_x} (the observed data matrix where each row is a cell and each column is one of \code{nrow(df_x)} variables),
\code{obs_y} (the observed data matrix where each row is a cell and each column is one of \code{nrow(df_y)} variables),
\code{true_x} (the true data matrix, which is \code{obs_x} without the technical noise),
\code{true_y} (the true data matrix, which is \code{obs_y} without the technical noise), and
\code{df_info} (the data frame with the same number of rows as \code{obs_x} and \code{obs_y}, where each row
contains meta-information about the cell)
}
\description{
Notes to self: 1) Currently, the function does not add technical noise.
2) It currently only generates data for Modality 1 as Bernoullis and for Modality 2 as Poissons.
}
