% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input.R
\name{generate_df_simple}
\alias{generate_df_simple}
\title{Generate a simple genome}
\usage{
generate_df_simple(
  p1,
  p2,
  genome_length = 10 * (p1 + p2),
  window = max(floor(genome_length/(4 * p2)), 1),
  intercept2 = 0,
  prefix1 = "peak",
  prefix2 = "gene"
)
}
\arguments{
\item{p1}{positive integer, number of variables in Modality 1}

\item{p2}{positive integer, number of variables in Modality 2}

\item{genome_length}{positive integer, for length of the genome}

\item{window}{positive integer, for the length of each the \code{p2} variables "shoulder" 
where the \code{round(p2/p1)} variables are located}

\item{intercept2}{numeric, for baseline signal for each of the \code{p2} variables}

\item{prefix1}{string, for the prefix for the \code{p1} variable names}

\item{prefix2}{string, for the prefix for the \code{p2} variable names}
}
\value{
two data frames, \code{df_x} and \code{df_y}, having \code{p1} and \code{p2}
rows for information of their respective variables
}
\description{
Generate a simple genome
}
