% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input.R
\name{generate_gcoef_simple}
\alias{generate_gcoef_simple}
\title{Generate coefficients forming the deterministic
relation from Modality 1 to Modality 2}
\usage{
generate_gcoef_simple(
  df_x,
  df_y,
  window = max(floor(0.5 * (nrow(df_y)/nrow(df_x))), 1),
  sparsity = 0,
  signal_mean = 3 * nrow(df_y)/nrow(df_x),
  signal_sd = 0
)
}
\arguments{
\item{df_x}{data frame where each row contains information about each variable in Modality 1}

\item{df_y}{data frame where each row contains information about each variable in Modality 2}

\item{window}{positive integer, for the genomic window that variables in Modality 1
can affect variables in Modality 2.}

\item{sparsity}{positive value between 0 and 1, denoting the percentage of 
variables in Modality 2 in within \code{window} of the variable in Modality 1 are "turned off"}

\item{signal_mean}{mean value of each non-zero value}

\item{signal_sd}{standard deviation of each non-zero value}
}
\value{
matrix with \code{nrow(df_x)} rows and \code{nrow(df_y)} columns
}
\description{
The data frames \code{df_x} and \code{df_y} can be from \code{generate_df_simple}.
The input \code{window} below should ideally that used in \code{generate_df_simple}.
}
