% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input.R
\name{generate_traj_cascading}
\alias{generate_traj_cascading}
\title{Generate cascading-from-the-left pattern}
\usage{
generate_traj_cascading(
  df_mod,
  steepness = 10,
  start_midpoint = 0,
  end_midpoint = 1,
  timepoints = 10 * nrow(df_mod),
  max_val = 1,
  min_val = 0
)
}
\arguments{
\item{df_mod}{data frame, for example, either \code{df_x} or \code{df_y} used as
input in \code{generate_gcoef_simple}.}

\item{steepness}{positive numeric denoting the steepness of the logistic function
(when the domain is from 0 to 1)}

\item{start_midpoint}{numeric denoting where the logistic function's mid starts off}

\item{end_midpoint}{numeric denoting where the logistic function's mid ends up}

\item{timepoints}{discretization of how many rows are outputted by this function.
The larger this number is, the higher the resolution and the more rows there are}

\item{max_val}{maximum value of the output matrix (formed by rescaling the logistic function)}

\item{min_val}{minimum value of the output matrix (formed by rescaling the logistic function)}
}
\value{
matrix with \code{timepoints} rows and \code{nrow(df_mod)} columns
}
\description{
Generate cascading-from-the-left pattern
}
