% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_cv.R
\name{lineage_cv}
\alias{lineage_cv}
\title{Lineage imputation with cross validation}
\usage{
lineage_cv(
  cell_features,
  cell_lineage,
  future_timepoint,
  lineage_future_count,
  lambda_initial,
  lambda_sequence_length,
  tab_mat,
  num_folds = 10,
  savefile_tmp = NULL,
  seed_number = 10,
  verbose = 0
)
}
\arguments{
\item{cell_features}{A numeric matrix where each row represents a cell, and each 
column represents a feature (for instance, the fastTopics scores). 
Let \code{n} denote the number of cells (rows). 
Please ensure there are row names for \code{cell_features} (denoting the 
cell IDs), and column names (denoting the feature names).}

\item{cell_lineage}{A character or factor vector of length \code{n} where 
element \code{i} of \code{cell_lineage} denotes which lineage cell \code{i}
belongs to.}

\item{future_timepoint}{A character that is one of the column names
of \code{tab_mat} to denote which column denotes which timepoint
is the "future" time point .}

\item{lineage_future_count}{A named numeric vector (where the names are the
lineage names that appeared in \code{cell_lineage}) that denotes the
number of cells at the future timepoint for each lineage.}

\item{lambda_initial}{The initial value of lambda to perform cross-validation on.}

\item{lambda_sequence_length}{The number of lambdas to perform cross-validation on.
The search starts with \code{lambda_initial} and then decays exponentially to 0.}

\item{tab_mat}{A matrix where the rows are named and are of each lineage
(which appeared in \code{cell_lineage}). There are two columns, one for
the number of cells in the current timepoint (corresponding to \code{cell_lineage}).
The other is the number of cells in the future timepoint (corresponding to
\code{lineage_future_count}).}

\item{num_folds}{Number of folds to do cross-validation on. Default is \code{10}.}

\item{savefile_tmp}{Filepath to save files to. Default is \code{NULL} (no 
temporary save files).}

\item{seed_number}{Seed value for reproducibility reasons. Default is \code{10}.}

\item{verbose}{A numeric, where numbers larger than 1 successively request more 
information to be printed out as the algorithm proceeds.}
}
\value{
A list with the number of elements corresponding to \code{num_folds}.
Each element of this list contains: \code{test_loglik} (the negative log-likelihood
on the held-out lineages), \code{train_loglik} (the negative log-likelihood on the trained
lineages), and \code{train_fit} (the actual fit, after using the \code{lineage_imputation_sequence()}).
}
\description{
Runs K-fold CV over a decreasing sequence of \code{lambda} values produced by
\code{lineage_imputation_sequence()}, selecting \code{lambda} by held-out objective.
}
