% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_cv_finalize.R
\name{lineage_cv_finalize}
\alias{lineage_cv_finalize}
\title{Finalize lineage imputation after cross validation}
\usage{
lineage_cv_finalize(cell_features, cell_lineage, fit_res, lineage_future_count)
}
\arguments{
\item{cell_features}{A numeric matrix where each row represents a cell, and each 
column represents a feature (for instance, the fastTopics scores). 
Let \code{n} denote the number of cells (rows). 
Please ensure there are row names for \code{cell_features} (denoting the 
cell IDs), and column names (denoting the feature names).}

\item{cell_lineage}{A character or factor vector of length \code{n} where 
element \code{i} of \code{cell_lineage} denotes which lineage cell \code{i}
belongs to.}

\item{fit_res}{This is the output of \code{lineage_cv()}.}

\item{lineage_future_count}{A named numeric vector (where the names are the
lineage names that appeared in \code{cell_lineage}) that denotes the
number of cells at the future timepoint for each lineage.}
}
\value{
A list with the following elements: \code{cell_imputed_score} (a
vector of length \code{nrow(cell_features)}) that denotes the predicted
progenies spawning from each particular cell, \code{coefficient_vec} (the
coefficient vector of lenght \code{ncol(cell_features)+1}) that denotes
the coefficients in the GLM, \code{lambda} (the chosen parameter after
cross-validation), and \code{lineage_imputed_count} (the vector of length
\code{lineage_future_count} that denotes the number of predicted cells 
at the future timepoint in each lineage).
}
\description{
This function is used after running \code{multiomeFate::lineage_cv()}.
Chooses \code{lambda} by minimizing the median held-out objective across folds,
then refits once on all cells at the chosen \code{lambda}.
}
