% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{nearest_neighbor}
\alias{nearest_neighbor}
\title{Construct nearest-neighbor object}
\usage{
nearest_neighbor(mat, nn_options)
}
\arguments{
\item{mat}{full data for a modality, where each row is a cell and each column is a variable}

\item{nn_options}{one of the outputs from \code{.chrom_options}}
}
\value{
exposed C++ object
}
\description{
Returns the exposed C++ object containing the nearest-neighbor
information (via the \code{RcppAnnoy} package) so
other methods later can query said object to find the nearest
points to said query
}
