% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_activation}
\alias{plot_activation}
\title{Plot activation}
\usage{
plot_activation(dat, mode = 1, cutoff = 0.5, reorder = T, ...)
}
\arguments{
\item{dat}{object of class \code{mf_simul}, from the output of \code{generate_data}}

\item{mode}{\code{1} or \code{2}, denoting whether or not the plot is of Modality 1 or 2 respecitvely}

\item{cutoff}{numeric, for the value that needs to be exceeded for a point to be plotted}

\item{reorder}{boolean. If \code{TRUE}, order the cells by their pseudotime.
If \code{FALSE}, order the cells by the order they appear in simulation}

\item{...}{additional graphical parameters}
}
\value{
nothing. A plot is made
}
\description{
Plot activation
}
