% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_arrow_iteration}
\alias{plot_arrow_iteration}
\title{Plot pseudotime direction over recruitment index}
\usage{
plot_arrow_iteration(res, vec_time, jitter = T, arrow_length = 0.1, ...)
}
\arguments{
\item{res}{object of class \code{chromatin_potential}, from the output of \code{chromatin_potential}}

\item{vec_time}{numeric vector of pseudotimes, with length \code{nrow(res$mat_x)}}

\item{jitter}{boolean}

\item{arrow_length}{numeric}

\item{...}{additional graphical parameters}
}
\value{
nothing. A plot is made
}
\description{
Plot pseudotime direction over recruitment index
}
