% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap.chromatin_potential}
\alias{plot_umap.chromatin_potential}
\title{Plot UMAP embedding (for \code{chromatin_potential})}
\usage{
\method{plot_umap}{chromatin_potential}(
  obj,
  mode_x = T,
  mode_y = T,
  k = 10,
  percent_arrows = 0.3,
  arrow_length = 0.05,
  col_vec = NA,
  ...
)
}
\arguments{
\item{obj}{object of class \code{chromatin_potential}, from the output of \code{chromatin_potential}}

\item{mode_x}{boolean, where \code{TRUE} means the data from Modality 1 is involved in the UMAP}

\item{mode_y}{boolean, where \code{TRUE} means the data from Modality 2 is involved in the UMAP}

\item{k}{positive numeric for the number of principal scores extracted}

\item{percent_arrows}{numeric}

\item{arrow_length}{numeric}

\item{col_vec}{\code{NA} or vector of colors}

\item{...}{additional graphical parameters}
}
\value{
nothing. A plot is made
}
\description{
Plot UMAP embedding (for \code{chromatin_potential})
}
