% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP embedding}
\usage{
plot_umap(
  dat,
  mode_x = T,
  mode_y = T,
  noiseless = T,
  k = 10,
  reorder = T,
  num_col = 10,
  ...
)
}
\arguments{
\item{dat}{object of class \code{mf_simul}, from the output of \code{generate_data}}

\item{mode_x}{boolean, where \code{TRUE} means the data from Modality 1 is involved in the UMAP}

\item{mode_y}{boolean, where \code{TRUE} means the data from Modality 2 is involved in the UMAP}

\item{noiseless}{boolean, where \code{TRUE} means information from 
\code{dat$true_x} and/or \code{dat$true_y} is plotted, and \code{FALSE}
means information from \code{dat$obs_x} and/or \code{dat$obs_y} is plotted}

\item{k}{positive numeric for the number of principal scores extracted}

\item{reorder}{boolean. If \code{TRUE}, color the cells by their pseudotime.
If \code{FALSE}, ccolor the cells by the order they appear in simulation}

\item{num_col}{positive numeric, for the number of distinct colors used}

\item{...}{additional graphical parameters}
}
\value{
nothing. A plot is made
}
\description{
Plot UMAP embedding
}
