% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_obj_nextcell.R
\name{prepare_obj_nextcell}
\alias{prepare_obj_nextcell}
\title{Generate object to prepare for simulating data}
\usage{
prepare_obj_nextcell(
  df_x,
  df_y,
  mat_g,
  list_traj_mat,
  branching_graph = NA,
  coarseness = 0.1,
  max_y = 1e+05,
  verbose = T
)
}
\arguments{
\item{df_x}{data frame where each row contains information about each variable in Modality 1}

\item{df_y}{data frame where each row contains information about each variable in Modality 2}

\item{mat_g}{matrix with \code{nrow(df_x)} rows and \code{nrow(df_y)} columns containing the linear coefficients on how
Modality 1's value affects the mean of Modality 2's values (i.e., the deterministic
relation from Modality 1 to Modality 2)}

\item{list_traj_mat}{list of matrices, where each matrix contains either \code{nrow(df_x)}
or \code{nrow(df_y)} columns. Each element of the list represents a different
segment in the trajectories. Each of the matrices within the list must have
column names matching either \code{df_x$name} or \code{df_y$name}.}

\item{branching_graph}{\code{igraph} object describing the branching structure}

\item{coarseness}{number between 0 and 1, describing how many rows in \code{list_traj_mat}
are used to fit the probabilistic relation from Modality 2 to Modality 1. The larger the number,
the more coarse (i.e., more approximate) the probabilistic relation is, but potentially
computationally more stable.}

\item{max_y}{positive integer, denoting the maximum value in Modality 2}

\item{verbose}{boolean}
}
\value{
object of class \code{mf_obj_next} containing the following components:
\code{df_x} and \code{df_y} and \code{mat_g} (all same as input), 
\code{ht} (a \code{hash} object storing all the pseudotimes and logistic regression coefficient
comprising the probabilistic relation from Modality 2 to Modality 1),
\code{mat_y2all} (a matrix with number of rows equal to \code{length(ht)},
where row \code{i} in this matrix is the Modality 2 vector corresponding to
the information in \code{ht[[as.character(i)]]}, and number of columns equal to
\code{nrow(df_y)})
\code{vec_startx} (the initial value for the cell in Modality 1),
\code{vec_starty} (the initial value for the cell in Modality 2)
}
\description{
Additional requirements: \code{df_x} and \code{df_y} are required to be data frames
with a column named \code{name}. The output of this function is meant as the input for
\code{generate_data}
}
\details{
Notes to self: 1) All entries in \code{mat_g} currently need to be non-negative,
2) Only linear trajectories are allowed currently, 3) Modality 1 (represented by
\code{df_x}) is assumed be Bernoulli and Modality 2 (represented by \code{df_y})
is assumed by Poisson. 4) Currently, the order of names in \code{mat_g} and 
\code{list_traj_mat} must match the exact order of names in \code{df_x} and \code{df_y}
}
