% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatin_potential.R
\name{chromatin_potential}
\alias{chromatin_potential}
\title{Main function to estimate chromatin potential}
\usage{
chromatin_potential(
  prep_obj,
  mat_g_init = NA,
  vec_g_init = rep(0, ncol(mat_y)),
  vec_threshold_init = rep(0, ncol(mat_y)),
  df_cell = NA,
  bool_oracle = F,
  verbose = T
)
}
\arguments{
\item{prep_obj}{object of class \code{chromatin_potential_prep}, created using the
\code{chromatin_potential_prepare} function}

\item{mat_g_init}{(optional) initial estimate of the matrix of coefficients linking Modality 1 to Modality 2}

\item{vec_g_init}{(optional) initial estimate of the vector of intercepts linking Modality 1 to Modality 2}

\item{vec_threshold_init}{(optional) initial estimate of the vector of thresholds}

\item{df_cell}{(optional) data frame containing information of each cell, primarily used
for simulations where the true cell types are known (for \code{.recruit_next})}

\item{bool_oracle}{boolean, where \code{TRUE} is used if you want to invoke "oracle" behavior
where true psuedotimes and cell-types might be known}

\item{verbose}{boolean}
}
\value{
object of class \code{chromatin_potential}
}
\description{
\code{.init_est_matrices} forms 2 different matrices
that this function updates every iteration (via \code{.update_estimation_matrices}):
\itemize{
\item \code{mat_x1}: This is the matrix about Modality 1
(i.e., has \code{ncol(mat_x)} variables)
where each row is a cell that has been previously-recruited, grabbed
from \code{mat_x}
\item \code{mat_y2}: This is the matrix about Modality 2
(i.e., has \code{ncol(mat_y)} variables)
where each row is the "future" cell matched
to the corresponding cell in \code{mat_x1}, grabbed
from \code{mat_y}. That is, this matrix has the same number of 
rows as \code{mat_x1}, but the \code{i}th row in
\code{mat_x1} might represent a different cell than the
\code{i}th row in \code{mat_y2}
}
In short, \code{mat_x1} and \code{mat_y2} always have the same number
of rows (but might represent different cells), and these two matrices
are used for \code{.estimate_g} to estimate the link from Modality 1 
to Modality 2.
}
