% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatin_potential.R
\name{chromatin_potential}
\alias{chromatin_potential}
\title{Main function to estimate chromatin potential}
\usage{
chromatin_potential(
  mat_x,
  mat_y,
  df_x,
  df_y,
  vec_start,
  list_end,
  form_method = "average",
  est_method = "glmnet",
  cand_method = "nn_xonly_avg",
  rec_method = "nn_yonly",
  options = list(),
  verbose = T
)
}
\arguments{
\item{mat_x}{full data for Modality 1, where each row is a cell and each column is a variable}

\item{mat_y}{full data for Modality 2, where each row is a cell and each column is a variable}

\item{df_x}{the data frame containing information of Modality 1}

\item{df_y}{the data frame containing information of Modality 2}

\item{vec_start}{integers between 1 and \code{nrow(mat_x)} to denote the cells at the start state}

\item{list_end}{list of integers between 1 and \code{nrow(mat_x)} to denote the cells any of the end states}

\item{form_method}{string}

\item{est_method}{string}

\item{cand_method}{string}

\item{rec_method}{string}

\item{options}{list}

\item{verbose}{boolean}
}
\value{
object of class \code{chromatin_potential}
}
\description{
\code{.init_est_matrices} forms 2 different matrices
that this function updates every iteration (via \code{.update_estimation_matrices}):
\itemize{
\item \code{mat_x1}: This is the matrix about Modality 1
(i.e., has \code{ncol(mat_x)} variables)
where each row is a cell that has been previously-recruited, grabbed
from \code{mat_x}
\item \code{mat_y2}: This is the matrix about Modality 2
(i.e., has \code{ncol(mat_y)} variables)
where each row is the "future" cell matched
to the corresponding cell in \code{mat_x1}, grabbed
from \code{mat_y}. That is, this matrix has the same number of 
rows as \code{mat_x1}, but the \code{i}th row in
\code{mat_x1} might represent a different cell than the
\code{i}th row in \code{mat_y2}
}
In short, \code{mat_x1} and \code{mat_y2} always have the same number
of rows (but might represent different cells), and these two matrices
are used for \code{.estimate_g} to estimate the link from Modality 1 
to Modality 2.
}
