% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_method.R
\name{.candidate_set}
\alias{.candidate_set}
\title{Find candidate cells}
\usage{
.candidate_set(mat_x, mat_y, df_res, cand_options)
}
\arguments{
\item{mat_x}{full data for Modality 1, where each row is a cell and each column is a variable}

\item{mat_y}{full data for Modality 2, where each row is a cell and each column is a variable}

\item{df_res}{data frame recording the current results, generated within \code{chromatin_potential}}

\item{cand_options}{one of the outputs from \code{.chrom_options}}
}
\value{
a vector of integers between 1 and \code{nrow(mat_x)}
}
\description{
Based on the cells previously recruited (recorded in \code{df_res}),
find other cells based on \code{mat_x} (the data for Modality 1) and
\code{df_res} to potentially be recruited in this current
iteration.
}
\details{
The options are:
\itemize{
\item \code{nn_xonly_any}: Find the candidates based only on which
unrecruited cells are one of the \code{cand_options$nn} nearest neighbors
to any previously-recruited cells in Modality 1
\item \code{nn_xonly_avg}: Find the \code{cand_options$num_cand} candidates 
which are \code{cand_options$num_cand} unrecruited cells which have the closest
\code{cand_options$nn} nearest neighbors of previously-recruited cells in Modality 1,
where we compute the average distance to said nearest neighbors using
\code{cand_options$average}
\item \code{all}: All the unrecruited cells are considered as candidates
}
}
