% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatin_potential_utils.R
\name{.chrom_options}
\alias{.chrom_options}
\title{Check all the options for chromatin potential}
\usage{
.chrom_options(form_method, est_method, cand_method, rec_method, options)
}
\arguments{
\item{form_method}{string}

\item{est_method}{string}

\item{cand_method}{string}

\item{rec_method}{string}

\item{options}{list}
}
\value{
list of 4 lists
}
\description{
This function checks all that options in \code{options} are appropriate
for the method specified in \code{form_method}, \code{est_method}, \code{cand_method}
and \code{rec_method},
and fills in the default values for any options not present in \code{options}
}
