% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_method.R
\name{.recruit_next}
\alias{.recruit_next}
\title{Based on the candidate cells, recruit a specific cell}
\usage{
.recruit_next(mat_x, vec_cand, mat_y1, idx1, res_g, df_res, rec_options)
}
\arguments{
\item{mat_x}{full data for Modality 1, where each row is a cell and each column is a variable}

\item{vec_cand}{output of \code{.candidate_set}}

\item{mat_y1}{output of \code{.init_est_matrices} or \code{.update_estimation_matrices}, representing
the data for Modality 1}

\item{idx1}{output of \code{.init_est_matrices} or \code{.update_estimation_matrices}, representing
the data for Modality 1}

\item{res_g}{output of \code{.estimate_g}}

\item{df_res}{data frame recording the current results, generated within \code{chromatin_potential}}

\item{rec_options}{one of the outputs from \code{.chrom_options}}
}
\value{
a list of vector of integers \code{vec_from} and a list
of integers \code{list_to}
}
\description{
Here, \code{mat_x} are all the cells in our dataset, among which
we are saying that \code{idx1} are cells that have previously been
recruited already, and \code{vec_cand} are the currently-selected
candidates that we are hoping to recruit and match to one of the
previously-recruited cells in \code{idx1}. This 
matching is done by using our estimation function \code{res_g}
on the candidate cell's Modality 1 expression in \code{mat_x}
and seeing if it's close to any of the Modality 2 expressions
in \code{mat_y1}.
}
\details{
The options are:
\itemize{
\item \code{nn_yonly}: Recruit \code{rec_options$num_rec} cells
whose predicted Modality 2 expression has the smallest average
(for example, mean, or in general, depending on what \code{rec_options$average}
is set to) distance to its \code{rec_options$nn} previously-recruited
nearest neighbors based on \code{mat_y1}
}
}
