% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_imputation_sequence.R
\name{lineage_imputation_sequence}
\alias{lineage_imputation_sequence}
\title{Fit lineage imputation along a lambda path}
\usage{
lineage_imputation_sequence(
  cell_features,
  cell_lineage,
  lineage_future_count,
  lambda_initial = NA,
  lambda_max = 101,
  lambda_min = 101,
  lambda_sequence_length = 50,
  multipler = 10000,
  verbose = 1
)
}
\arguments{
\item{cell_features}{A numeric matrix where each row represents a cell, and each 
column represents a feature (for instance, the fastTopics scores). 
Let \code{n} denote the number of cells (rows). 
Please ensure there are row names for \code{cell_features} (denoting the 
cell IDs), and column names (denoting the feature names).}

\item{cell_lineage}{A character or factor vector of length \code{n} where 
element \code{i} of \code{cell_lineage} denotes which lineage cell \code{i}
belongs to.}

\item{lineage_future_count}{A named numeric vector (where the names are the
lineage names that appeared in \code{cell_lineage}) that denotes the
number of cells at the future timepoint for each lineage.}

\item{lambda_initial}{The initial value of lambda to perform cross-validation on.}

\item{lambda_max, lambda_min}{Bounds used when computing an internal \code{lambda_initial}.}

\item{lambda_sequence_length}{The number of lambdas to perform cross-validation on.
The search starts with \code{lambda_initial} and then decays exponentially to 0.}

\item{multipler}{Scaling factor for the internal \code{lambda_initial} heuristic.}

\item{verbose}{A numeric, where numbers larger than 1 successively request more 
information to be printed out as the algorithm proceeds.}
}
\value{
A list with \code{fit_list} (solution estimated by \code{lineage_imputation()} per lambda) and \code{lambda_sequence}.
}
\description{
This function calls \code{lineage_imputation()} for a sequence of lambdas.
}
