% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP embedding (generic)}
\usage{
plot_umap(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{additional parameters}
}
\value{
nothing. A plot is made
}
\description{
Plot UMAP embedding (generic)
}
